feat(core/*): introduce strapi.ai namespace (#25886)
