+++ packages/core/admin/admin/src/ee.ts
@@ -4,5 +4,5 @@
 
 export { useLicenseLimits } from '../../ee/admin/src/hooks/useLicenseLimits';
 export { useEnterprise } from './hooks/useEnterprise';
-export { useGetAIUsageQuery, useGetAIFeatureConfigQuery } from '../../ee/admin/src/services/ai';
+export { useGetAiUsageQuery, useGetAiFeatureConfigQuery } from '../../ee/admin/src/services/ai';
 export { useAIAvailability } from '../../ee/admin/src/hooks/useAIAvailability';
