+++ packages/core/content-type-builder/admin/src/components/AIChat/hooks/useAIFetch.ts
@@ -8,7 +8,7 @@ import { useState } from 'react';
 
 import { UIMessage, useChat } from '@ai-sdk/react';
 import { useAppInfo } from '@strapi/admin/strapi-admin';
-import { useGetAIUsageQuery } from '@strapi/admin/strapi-admin/ee';
+import { useGetAiUsageQuery } from '@strapi/admin/strapi-admin/ee';
 import { DefaultChatTransport } from 'ai';
 
 import { fetchAI, makeChatFetch, safeParseJson } from '../lib/aiClient';
@@ -149,7 +149,7 @@ export const createAIFetchHook = <T extends keyof AIEndpoints>(endpoint: T) => {
     const strapiVersion = useAppInfo('useAIFetch', (state) => state.strapiVersion);
     const projectId = useAppInfo('useAIFetch', (state) => state.projectId);
     const userId = useAppInfo('useAIFetch-user', (state) => state.userId);
-    const aiUsage = useGetAIUsageQuery(undefined, { refetchOnMountOrArgChange: true });
+    const aiUsage = useGetAiUsageQuery(undefined, { refetchOnMountOrArgChange: true });
 
     const [isPending, setIsPending] = useState(false);
     const [error, setError] = useState<string | null>(null);
