+++ packages/core/core/src/providers/index.ts
@@ -1,4 +1,5 @@
 import admin from './admin';
+import ai from './ai';
 import coreStore from './coreStore';
 import cron from './cron';
 import registries from './registries';
@@ -11,6 +12,7 @@ import type { Provider } from './provider';
 export const providers: Provider[] = [
   registries,
   admin,
+  ai,
   coreStore,
   sessionManager,
   webhooks,
