+++ packages/core/upload/server/src/services/metrics.ts
@@ -7,14 +7,15 @@ const isProviderPrivate = async () => strapi.plugin('upload').provider.isPrivate
 export default ({ strapi }: { strapi: Core.Strapi }) => ({
   async trackUsage(event: string, properties?: Record<string, any>) {
     const settings: Settings = await strapi.plugin('upload').service('upload').getSettings();
-    const isAIAvailable =
-      strapi.config.get('admin.ai.enabled', true) && strapi.ee.features.isEnabled('cms-ai');
+    const isAiAvailable = strapi.ai.admin.isEnabled();
 
     return strapi.telemetry.send(event, {
       ...properties,
       eventProperties: {
         ...properties?.eventProperties,
-        ...(isAIAvailable ? { isAIMediaLibraryConfigured: Boolean(settings.aiMetadata) } : {}),
+        ...(isAiAvailable === true
+          ? { isAiMediaLibraryConfigured: Boolean(settings.aiMetadata) }
+          : {}),
       },
     });
   },
