+++ packages/plugins/i18n/admin/src/services/settings.ts
@@ -14,7 +14,7 @@ const settingsApi = i18nApi.injectEndpoints({
         method: 'PUT',
         data,
       }),
-      invalidatesTags: ['Settings', 'AIFeatureConfig'],
+      invalidatesTags: ['Settings', 'AiFeatureConfig'],
     }),
   }),
 });
